html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/creating_browsing_context_test_01.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>[Browsing Context] : [APIs for creating browsing_contexts by name]</title>
<link rel="author" title="Duhyeong Kim" href="mailto:dduskim@gmail.com">
<link rel="help" href="http://www.w3.org/TR/2009/WD-html5-20090423/browsers.html#apis-for-creating-and-navigating-browsing-contexts-by-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<script>
	//	arguments test
	var test1 = async_test('creating browsing context first argument test 1');
	var test2 = async_test('creating browsing context first argument test 2');

	var currentUrl = 'http://' + window.location.host + '/';
	var win1 = window.open(currentUrl, '', 'height=1,width=1');

	setTimeout(function () {
		test1.step(function () {
			assert_equals(win1.window.location.href, currentUrl, 'should be equal to result url');
			win1.close();
		});
		test1.done();
	}, 500);

	var win2 = window.open('', '', 'height=1,width=1');
	setTimeout(function () {
		test2.step(function () {
			assert_equals(win2.window.location.href, 'about:blank', 'should be equal to "about:blank"');
			win2.close();
		});
		test2.done();
	}, 500);

	var win3 = window.open('', 'testWindow', 'height=1,width=1');
	test(function () {
		win3.close();
		assert_equals(win3.name, 'testWindow', 'should have a browser context name');
	}, 'creating browsing context second argument test 3');

	var win4 = window.open('', '', 'height=1,width=1');
	test(function () {
		assert_equals(win4.name, '', 'window should have a name');
		win4.name = 'testWindow';
		assert_equals(win4.name, 'testWindow', 'should be equal to ""');
		win4.close();
	}, 'creating browsing context second argument test 4');

</script>
